Add support for hi-res versions of images for PDF
The ◊image, ◊margin-figure and ◊figure tags now also accept an optional
boolean parameter `#:has-print-version`. If this is #t, the tag will
specify a different folder for the image source (the image filename
must be the same).

This allows you to use a hi-res copy of (say) a photo in PDF output,
and use a smaller/optimized version of the same photo on the the web
page.