Tidy HTML output
Racket’s HTML output can be hard to read (not that it really matters).
Racket has a way to “pretty print“ HTML/XML (display-xml) but it is
opinionated and breaks validation. Using an separate, purpose-built
utility like `tidy` is a good way to clean it all up afterwards.

Adding `|| true` ensures that any warnings or error codes returned by
tidy are suppressed, so they don’t break the build.