Moving from boot to deps.edn
Using the ShimDandy environment for ash-ra-template means abandoning the
abstractions of Perun and Boot because they don't play nicely with the Pod
concept.

I don't need so many of Perun's features anyway. I can probably replicate a lot
of the page collection features with (file-seq) and build my own tools.