[HOLOTYPE] 4.2 / benchmark performance slowdowns in dtype.next
Reflection appears to be the culprit, as it typically is,
but the nREPL/transit layer appears to be a confounding factor.

Type hints may eliminate this performance penalty, but in order to
leverage the type information I may have to rewrite the pipeline
operation to not use the thread macro.